requirements.txt
google-auth>=2.38.0
google-api-python-client>=2.161.0
google-auth-oauthlib>=1.2.1
openai>=1.0.0
python-dateutil>=2.8.2
python-dotenv>=0.19.0
pytz>=2024.1
spacy>=3.7.4
holidays>=0.45
sqlalchemy>=2.0.0
alembic>=1.13.0
click>=8.1.7
rich>=13.7.0
prompt_toolkit>=3.0.43
keyring>=24.3.0
fastapi>=0.109.2
uvicorn>=0.27.1
websockets>=12.0
